+++ R/mplus.update.R
@@ -671,7 +671,7 @@ mplus.update <- function(x, update, file = "Mplus_Input_Update.inp", replace.inp
                              variable = variable, not.input = not.input, not.result = not.result,
                              write = write, append = append, check = check, output = output),
                  input = input.object, write = write.object,
-                 result = result.object)
+                 result = result.object$result)
 
   class(object) <- "misty.object"